Add new components/Admin directory and move CreateRewards components under there, where they belong
